added a checkpoint system
I attempted a large scrape and there was an error halfway through. All data was lost and will need to be scraped again. At least with a checkpoint in place not all data will be lost and theres a point to start from next time.

Also added a sleep.timer to be more courteous to the server